+++ build-scripts/builder/entrypoint.sh
@@ -118,7 +118,6 @@ download_snapshots() {
     echo "----------------------------------------"
 
     bash ~/build-scripts/download_snapshots.sh \
-        --env "https://api.pre.cloud.wazuh.com/api/v1" \
         --output-dir ~/artifacts/snapshots
 }